Fixed __setitem__ failure for pandas series
Pysat checks for an iterable of DataFrames by checking the type of
element 0 using [0], but not all iterables have a [0], the more
universal way of checking is to cast to an iterator and use the next
function. An empty iterator is set to return None to avoid throwing a
StopIteration.